BigDecimal summa

+++ src/main/java/ru/division/of/expenses/app/models/DirectPayer.java
@@ -14,7 +14,7 @@ public class DirectPayer extends AbstractEntity {
     private User user;
 
     @Column
-    private Double summa;
+    private BigDecimal summa;
 
     @ManyToOne
     @JoinColumn(name = "expense_id")
